day 11: part 1 implementation
Unfortunately, this program cannot compile and determine the actual number of nodes, as COBOL has a numerical limit of 2147483647, and no string can be longer than 268435456. Therefore, a COBOL program alone cannot be used to solve part 2 of this problem. Thus, this program cannot actually compile. If it could, this is what it would look like.